continue running when querying the jenkins api fails
Instead of terminating when the Jenkins Server can not be reachted, the jenkins
exporter continues to run.
This prevents that the jenkins-exporter terminates when e.g. the jenkins-server
is restarted.

A new Prometheus metric jenkins_exporter_errors is added, that counts the number
of encountered errors when trying to query the jenkins api.