change metric type to histogram
Instead of exporting the build duration as summary, export them as histogram.
The summary metric has the issue that it calculates the quantile on the client
over a sliding time window.
The time window was 10min (the default). This window is much too small to get
reasonable quantile values, it can happen that not a single build run for a
Jenkins job during the window.