remove old state file entries
Every 10min the jenkins-exporter now removes entries about the last seen build
ID of a job if the entry was not updated longer then maxStateAge.
maxStateAge can be set via a commandline flag, the default is 14days.

This solves the issue that jenkins job that do not exist anymore were kept in
the state file forever.