fix: go vet warning: sigchanyzer: misuse of unbuffered os.Signal channel
This fixes the go vet warning:
vet warning: sigchanyzer: misuse of unbuffered os.Signal channel as argument to
	signal.Notify (govet)

registerSigHandler() did start a go-routine that received from the channel
immediately after the signal notifier was registered.
Therefore a signal could be lost if it was received during startup, after the
signal.Notify() was called but before the go-routine was started.