only skip recording builds if state file did not exist
The jenkins exporter only recorded metrics for builds that happened after
the exporter saw a build of a job the first time.
The logic existed to prevent that metrics are recorded multiple times if the
state of a previous run was deleted.
It had the side-effect that the first build of multibranch job were never
recorded.
Improve this by only **not** recording builds if a new state file existed.
If a state file existed, assume that only the builds up to the recorded ones
were seen and record metrics for all unknown ones.

Also remove a wrong comment about if buildID of -1 should be recorded when a job
is seen that has no build. This is not needed. Jenkins starts counting builds at
1. If a jenkins job has no builds, the highest ID will be 0 and the first build
   will be recorded.