add -http-request-limit parameter
Add a -http-request-limit parameter that limits the number of http-requests sent
to Jenkins to 1 in the given interval.
This can be used to limit the load that the http-requests generate at the
jenkins server.
The polling interval is unaffected by -http-request-limit, the pause between
interval is always the same independent of how long it took (because of the rate
limit) to get all data for a poll interval.