support to record branch labels for multibranch jobs
Add a new -branch-label-allowlist cmdline parameter. It allows to specify
multibranch job and their branch names for which a "branch" metric label will be
recorded.
This allows to have different metrics per of a multibranch job.
If a branch is not in the allowlist the branch label will be empty.