refactor metric implementation
Replace the internal prometheus package with a struct holding the metric types
of the prometheus package.
This is necessary to be able to create metrics with all possible labels on
startup.
With the previous implementation the metric was created when a value was first
time recorded for it, which could be with or without a branch label.
When it is created without the branch label, it was not possible to add it for
other observations.
This also simplifies the implementation because we now simply rely on the
Prometheus package instead of implementing unnecessary wrappers.
The Metrics struct is stored in the new internal jenkinsexporter package.
The package could the whole functionality regarding fetching metrics from
jenkins and recording it that is currently implemented in main.
This can be refactored in a follow-up.
It works as it is for now. :-)