do not record runtime metrics for builds that are still running
To evaluate if a build is still running it is checked if build.Result is empty.
This is not sufficient, sometimes the result of a build is set but it is still
running. This causes that runtime metrics are recorded for builds that are still
in progress. The runtime metrics values are then too low.

To determine if a build is running fetch the building field of builds and
check if it is true instead.