do not skip metric recording for older builds that are in progress
When multiple builds run for the same job, it can happen that build N finishes
before build N-1. If an unfinished build is retrieved it was ignored.
If a build with a higher number already finished, it's
build number is recorded as highest seen number and all previous builds are
ignored.

This behavior caused that for a lot of builds no metrics were recorded which
screws the statistics.

Instead of ignoring unfinished build if a newer one already finished, it's build
number is now stored in a UnrecordedBuildIDs map.
If an following iteration the build finished and is in the UnrecordedBuildIDs
maps, metrics for it are recorded and it is removed from the map.

This required to change the format of the state file.
If a state file in the old format is found, jenkins-exporter will exit with an
error.
To upgrade, the previous state file must either be deleted or a path to a
different state file must be specified as parameter or via environment variable.