metrics: remove wrong jenkins_exporter_errors{type=http_requests_total} metric
The jenkins_http_requests_total and jenkins_exporter_errors metrics are mixed
up.
A metric called jenkins_exporter_errors with the label http_requests_total is
incremented for every http requests.
The name implied it only counts errors but this is not the case.
The documented metric called jenkins_http_requests_total is not recorded at all.

Remove jenkins_exporter_errors metrics with the http_requests_total label.
Remove mentions of the jenkins_http_requests_total metrics in the README.

If it turns out that there is a use case for the jenkins_http_requests_total
metric it can be added back.